% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grmforest.R
\name{grmforest}
\alias{grmforest}
\title{Fit a GRM Forest}
\usage{
grmforest(formula, data, control = grmforest.control(), ...)
}
\arguments{
\item{formula}{A formula specifying the model structure (response ~
predictors).}

\item{data}{A data frame containing the variables in the model.}

\item{control}{A control object created by \code{grmforest.control()}.}

\item{...}{Additional arguments passed to \code{grmtree()}.}
}
\value{
An object of class \code{grmforest} containing: \item{trees}{List of
fitted GRM trees} \item{oob_samples}{List of out-of-bag samples for each
tree} \item{formula}{The model formula} \item{data}{The original dataset}
\item{call}{The function call}
}
\description{
Fits a forest of graded response model trees using either bootstrap
aggregation (bagging) or subsampling. Each tree is fitted on a random sample
of the data, with out-of-bag samples retained for validation.
}
\examples{

\donttest{
  library(grmtree)
  library(hlt)
  data("asti", package = "hlt")
  asti$resp <- data.matrix(asti[, 1:4])

  # Fit forest with default parameters
  forest <- grmforest(resp ~ gender + group, data = asti)

  # Fit with custom control
  ctrl <- grmforest.control(n_tree = 50, sampling = "subsample")
  forest <- grmforest(resp ~ gender + group, data = asti, control = ctrl)
}

}
\seealso{
\code{\link{grmtree}} fits a Graded Response Model Tree,
\code{\link{grmtree.control}} creates a control object for \code{grmtree},
\code{\link{grmforest.control}} creates a control object for \code{grmforest},
\code{\link{varimp}} calculates the variable importance for GRM Forest,
\code{\link{plot.varimp}} creates a bar plot of variable importance scores
}
