% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grmforest.R
\name{print.grmforest}
\alias{print.grmforest}
\title{Print Method for GRM Forest}
\usage{
\method{print}{grmforest}(x, ...)
}
\arguments{
\item{x}{A \code{grmforest} object}

\item{...}{Additional arguments (currently unused)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for GRM Forest
}
